% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateND_committed.R
\name{FateND_committed}
\alias{FateND_committed}
\title{by inputing the fitting committed data and genepairs and all genes,we can get the cell fate determinates}
\usage{
FateND_committed(lineage1_fit, overall_genepair, U)
}
\arguments{
\item{lineage1_fit}{refer to the fitting  committed time series data}

\item{overall_genepair}{refer to all gene pairs}

\item{U}{refer to all genes}
}
\value{
cell fate determinate TFs
}
\description{
function to get cell fate determinate TFs
}
