% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateND_embryos.R
\name{FateND_embryos}
\alias{FateND_embryos}
\title{by inputing the fit data of two lineages and genepairs and all genes,we can get the cell fate determinates}
\usage{
FateND_embryos(lineage1_fit, lineage2_fit, overall_genepair, U)
}
\arguments{
\item{lineage1_fit}{refer to the fitting  time series data of lineage1 in embryos}

\item{lineage2_fit}{refer to the fitting  time series data of lineage2 in embryos}

\item{overall_genepair}{refer to all gene pairs}

\item{U}{refer to all genes}
}
\value{
cell fate determinate TFs
}
\description{
function to get cell fate determinate TFs
}
