% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateND_HSC.R
\name{FateND_HSC}
\alias{FateND_HSC}
\title{by inputing the fit data of three lineages and genepairs and all genes of HSC,we can get the cell fate determinates}
\usage{
FateND_HSC(lineage1_fit, lineage2_fit, lineage3_fit, overall_genepair, U)
}
\arguments{
\item{lineage1_fit}{refer to the fitting  time series data of lineage1 in HSC}

\item{lineage2_fit}{refer to the fitting  time series data of lineage2 in HSC}

\item{lineage3_fit}{refer to the fitting  time series data of lineage3 in HSC}

\item{overall_genepair}{refer to all gene pairs}

\item{U}{refer to all genes}
}
\value{
cell fate determinate TFs
}
\description{
function to get cell fate determinate TFs
}
