% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_committed.R
\name{get_fit_committed}
\alias{get_fit_committed}
\title{fit the time series committed data to the trend like switch genes and transient genes}
\usage{
get_fit_committed(lineage1)
}
\arguments{
\item{lineage1}{refer to time series of committed data}
}
\value{
g1 & g2 & fitdata
}
\description{
function to fit time series data by nls
}
