% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_embryos.R
\name{get_fit_embryos}
\alias{get_fit_embryos}
\title{fit the two lineages of time series embryos data to the trend like switch genes and transient genes}
\usage{
get_fit_embryos(lineage1, lineage2)
}
\arguments{
\item{lineage1}{refer to time series data of lineage1 in embryos}

\item{lineage2}{refer to time series data of lineage2 in embryos}
}
\value{
g1 & g2 & fitdata of lineage1 and lineage2
}
\description{
function to fit time series data by nls
}
