% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_HSC.R
\name{get_fit_HSC}
\alias{get_fit_HSC}
\title{fit the three lineages of time series HSC data to the trend like switch genes and transient genes}
\usage{
get_fit_HSC(lineage1, lineage2, lineage3)
}
\arguments{
\item{lineage1}{refer to time series data of lineage1 in HSC}

\item{lineage2}{refer to time series data of lineage2 in HSC}

\item{lineage3}{refer to time series data of lineage3 in HSC}
}
\value{
g1 & g2 & fitdata of lineage1 lineage2  and lineage3
}
\description{
function to fit time series data by nls
}
