% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genepair.R
\name{get_genepair}
\alias{get_genepair}
\title{get all genepairs by pairing the g1 and g2}
\usage{
get_genepair(A, B)
}
\arguments{
\item{A}{refer to g1}

\item{B}{refer to g2}
}
\value{
all gene pairs
}
\description{
function to get gene pair
}
